pyproject.toml
[tool.poetry]
name = "hw-24"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hw_24"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2 = "^2.9.7"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
